Add custom train API, it is working, but still need more love
At least we can call the train programatically. I haven't tested in the feature generators write the features in the right format (for deco they do work).
Now it should be nice to add (from another project!) a web service that makes use of this, to see from a client perspective which other necessities might arise.
Also, the features used for training are not registered anywhere (to be used later for deco). They should be stored in the same folder than the model, so the feature config can be read and applied at deco time.